requirements.txt
tensorflow
pandas
keras
matplotlib
numpy
scikit-learn
seaborn
scipy
torch
torchvision
nltk
pillow
tqdm
rouge-score
